requirements.txt
pandas
python-dotenv
numpy
scikit-learn
matplotlib
seaborn
pymongo
certifi
pymongo[srv]==3.12
dill
xgboost
catboost
flask


#-e.
